修改模型主键类型并创建几个用户系统相关模型类

+++ oneshop/src/main/java/com/gelvt/oneshop/mapper/BaseMapper.java
@@ -20,7 +20,7 @@ public interface BaseMapper<T extends Entity> {
 	/** 删除一个实体
 	 * @param id id
 	 */
-	void remove(Integer id);
+	void remove(Long id);
 	/** 更新一个实体
 	 * @param entity 待更新实体
 	 */
@@ -30,5 +30,5 @@ public interface BaseMapper<T extends Entity> {
 	 * @param id id
 	 * @return 持久化的实体
 	 */
-	T get(Integer id);
+	T get(Long id);
 }
